# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A page of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List users visible to the caller (own tenant; all for platform admin)",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a user (admins only; there is no self sign-up)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a user profile (admins may also activate or deactivate)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove_v1",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/roles": {
      "put": {
        "operationId": "UsersController_setRoles_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user with its new roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated, or the token is invalid or expired",
            "content": {
              "application/problem+json": {}
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to do this",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "No such resource, or the caller may not know it exists",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "The resource is in a state that forbids this change",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "The payload failed validation",
            "content": {
              "application/problem+json": {}
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace the roles of a user (never your own, never more than you hold)",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "Starter App API",
    "description": "Success responses are wrapped in `{ data, meta? }`; errors follow RFC 9457 Problem Details with `application/problem+json`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token only. Swagger adds the Bearer prefix."
      }
    },
    "schemas": {
...
      "UserListEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "fullName": {
                  "type": "string"
                },
                "phoneNumber": {
                  "type": "string",
                  "nullable": true
                },
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "key": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "key",
                      "name"
                    ]
                  }
                },
                "tenantId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "isActive": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "email",
                "fullName",
                "phoneNumber",
                "roles",
                "tenantId",
                "isActive",
                "createdAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 256
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[0-9]{7,15}$"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 256
          },
          "roleIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 20,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "fullName",
          "password",
          "roleIds"
        ]
      },
      "UserEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "fullName": {
                "type": "string"
              },
              "phoneNumber": {
                "type": "string",
                "nullable": true
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "name"
                  ]
                }
              },
              "tenantId": {
                "type": "string",
                "format": "uuid",
                "nullable": true
              },
              "isActive": {
                "type": "boolean"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "email",
              "fullName",
              "phoneNumber",
              "roles",
              "tenantId",
              "isActive",
              "createdAt"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[0-9]{7,15}$",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "SetUserRolesDto": {
        "type": "object",
        "properties": {
          "roleIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 20,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "roleIds"
        ]
      }
    }
  }
}